pyproject.toml
[project]
name = "TChat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=44.0.0",
    "pysocks>=1.7.1",
    "prompt-toolkit>=1.0.0",
]
